refactor: move CordovaError module up
