+++ src/CordovaError.js
